+++ accounts-db/src/rent_debits.rs
@@ -1,18 +1,8 @@
 use {
+    crate::stake_rewards::RewardInfo,
     solana_sdk::{pubkey::Pubkey, reward_type::RewardType},
     std::collections::HashMap,
 };
-
-#[derive(Debug, PartialEq, Eq, Serialize, Deserialize, AbiExample, Clone, Copy)]
-pub struct RewardInfo {
-    pub reward_type: RewardType,
-    /// Reward amount
-    pub lamports: i64,
-    /// Account balance in lamports after `lamports` was applied
-    pub post_balance: u64,
-    /// Vote account commission when the reward was credited, only present for voting and staking rewards
-    pub commission: Option<u8>,
-}
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct RentDebit {
