+++ geyser-plugin-manager/src/block_metadata_notifier_interface.rs
@@ -1,5 +1,5 @@
 use {
-    solana_accounts_db::rent_debits::RewardInfo,
+    solana_accounts_db::stake_rewards::RewardInfo,
     solana_sdk::{clock::UnixTimestamp, pubkey::Pubkey},
     std::sync::{Arc, RwLock},
 };
