+++ runtime/src/snapshot_package.rs
@@ -11,6 +11,7 @@ use {
         accounts_db::{AccountStorageEntry, IncludeSlotInHash, INCLUDE_SLOT_IN_HASH_TESTS},
         accounts_hash::{AccountsHash, AccountsHashEnum},
         epoch_accounts_hash::EpochAccountsHash,
+        rent_collector::RentCollector,
     },
     solana_sdk::{clock::Slot, feature_set, sysvar::epoch_schedule::EpochSchedule},
     std::{
@@ -22,7 +23,6 @@ use {
 
 mod compare;
 pub use compare::*;
-use solana_accounts_db::rent_collector::RentCollector;
 
 /// This struct packages up fields to send from AccountsBackgroundService to AccountsHashVerifier
 pub struct AccountsPackage {
